chore: move the utils/ into .vitepress/
