+++ docs/.vitepress/theme/Layout/Home.vue
@@ -1,14 +1,14 @@
 <script lang="ts">
 import { defineComponent } from "vue";
 import { useData, withBase } from "vitepress";
-import { data as posts } from "@@/.vitepress/data/routes.data.ts";
+import { data as posts } from "@@/data/routes.data.ts";
 import ArticleItem from "./components/ArticleItem.vue";
 export default defineComponent({
 	name: "Home",
 	components: { ArticleItem },
 	setup() {
-		const { site, frontmatter } = useData();
-		return { site, posts, withBase, frontmatter };
+		const { site } = useData();
+		return { site, posts, withBase };
 	},
 });
 </script>
@@ -17,10 +17,10 @@ export default defineComponent({
 	<div class="px-6">
 		<ul class="grid gap-4">
 			<ArticleItem
-				v-for="{
-					url,
-					frontmatter: { title, catagory, tags, date, last_updated: lastUpdated, og_url: ogUrl },
-				} in posts"
+				v-for="(
+					{ url, frontmatter: { title, catagory, tags, date, last_updated: lastUpdated, og_url: ogUrl } }, key
+				) in posts"
+				:key="key"
 				:title="title"
 				:url="url"
 				:catagory="catagory"
