+++ tsconfig.json
@@ -21,7 +21,7 @@
 		"noFallthroughCasesInSwitch": true,
 		"paths": {
 			"@/*": ["./src/*"],
-			"@@/*": ["./docs/*"]
+			"@@/*": ["./docs/.vitepress/*"]
 		},
 		"allowArbitraryExtensions": true
 	},
